stacking: finish Calibration->Stacking namespace split
The earlier split commit (9d027db) moved the files via git mv but the
namespace rewrite never got committed -- locally my working tree had
the fix so the build looked fine, but origin still declared
`namespace TianWen.Lib.Imaging.Calibration` inside `Stacking/*.cs`. CI
fell over with CS0246 on IntegrationResult / IntegrationProgress.

Rewrites the namespace header in all moved files and adds the
`using TianWen.Lib.Imaging.Stacking;` import to every consumer that
needed it (tests + bench + Calibration<->Stacking glue files).
