+++ src/TianWen.Lib.Tests/CalibratorTests.cs
@@ -2,6 +2,7 @@ using System;
 using Shouldly;
 using TianWen.Lib.Imaging;
 using TianWen.Lib.Imaging.Calibration;
+using TianWen.Lib.Imaging.Stacking;
 using Xunit;
 
 namespace TianWen.Lib.Tests;
