+++ src/TianWen.Lib.Tests/ChunkedTwoPassStrategyTests.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Shouldly;
 using TianWen.Lib.Imaging;
 using TianWen.Lib.Imaging.Calibration;
+using TianWen.Lib.Imaging.Stacking;
 using Xunit;
 
 namespace TianWen.Lib.Tests;
