+++ src/TianWen.Lib.Tests/FrameCacheTests.cs
@@ -1,5 +1,6 @@
 using Shouldly;
 using TianWen.Lib.Imaging.Calibration;
+using TianWen.Lib.Imaging.Stacking;
 using Xunit;
 
 namespace TianWen.Lib.Tests;
