+++ src/TianWen.Lib.Tests/IntegrationFitsWriterTests.cs
@@ -4,6 +4,7 @@ using System.IO;
 using Shouldly;
 using TianWen.Lib.Imaging;
 using TianWen.Lib.Imaging.Calibration;
+using TianWen.Lib.Imaging.Stacking;
 using Xunit;
 
 namespace TianWen.Lib.Tests;
