+++ src/TianWen.Lib.Tests/MasterFrameBuilderTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using Shouldly;
 using TianWen.Lib.Imaging;
 using TianWen.Lib.Imaging.Calibration;
+using TianWen.Lib.Imaging.Stacking;
 using Xunit;
 
 namespace TianWen.Lib.Tests;
