+++ src/TianWen.Lib.Tests/MemoryMappedFitsSinkTests.cs
@@ -3,6 +3,7 @@ using System.IO;
 using Shouldly;
 using TianWen.Lib.Imaging;
 using TianWen.Lib.Imaging.Calibration;
+using TianWen.Lib.Imaging.Stacking;
 using Xunit;
 
 namespace TianWen.Lib.Tests;
