+++ src/TianWen.Lib.Tests/MinMaxClipRejectorTests.cs
@@ -1,6 +1,7 @@
 using System;
 using Shouldly;
 using TianWen.Lib.Imaging.Calibration;
+using TianWen.Lib.Imaging.Stacking;
 using Xunit;
 
 namespace TianWen.Lib.Tests;
