+++ src/TianWen.Lib.Tests/NormalizerTests.cs
@@ -4,6 +4,7 @@ using System.Numerics;
 using Shouldly;
 using TianWen.Lib.Imaging;
 using TianWen.Lib.Imaging.Calibration;
+using TianWen.Lib.Imaging.Stacking;
 using Xunit;
 
 namespace TianWen.Lib.Tests;
