+++ src/TianWen.Lib.Tests/RegistrationSidecarTests.cs
@@ -4,6 +4,7 @@ using System.Numerics;
 using System.Threading.Tasks;
 using Shouldly;
 using TianWen.Lib.Imaging.Calibration;
+using TianWen.Lib.Imaging.Stacking;
 using Xunit;
 
 namespace TianWen.Lib.Tests;
