+++ src/TianWen.Lib.Tests/TilePipelinedStrategyTests.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using Shouldly;
 using TianWen.Lib.Imaging;
 using TianWen.Lib.Imaging.Calibration;
+using TianWen.Lib.Imaging.Stacking;
 using Xunit;
 
 namespace TianWen.Lib.Tests;
