+++ src/TianWen.Lib/Imaging/Stacking/ChunkedTwoPassStrategy.cs
@@ -2,7 +2,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace TianWen.Lib.Imaging.Calibration;
+namespace TianWen.Lib.Imaging.Stacking;
 
 /// <summary>
 /// Streaming two-pass strategy with cross-N rejection. Pass A accumulates
