+++ src/TianWen.Lib/Imaging/Stacking/Float16StagedStrategy.cs
@@ -4,7 +4,7 @@ using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace TianWen.Lib.Imaging.Calibration;
+namespace TianWen.Lib.Imaging.Stacking;
 
 /// <summary>
 /// Stage the full warped canvas as float16 -- half the disk bytes of float32
