+++ src/TianWen.Lib/Imaging/Stacking/FootprintStagedStrategy.cs
@@ -4,7 +4,7 @@ using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace TianWen.Lib.Imaging.Calibration;
+namespace TianWen.Lib.Imaging.Stacking;
 
 /// <summary>
 /// Stage each warped frame to disk as float32 but only over its non-NaN
