+++ src/TianWen.Lib/Imaging/Stacking/IIntegrationStrategy.cs
@@ -1,7 +1,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace TianWen.Lib.Imaging.Calibration;
+namespace TianWen.Lib.Imaging.Stacking;
 
 /// <summary>
 /// Identifies an <see cref="IIntegrationStrategy"/> implementation. Used by the
