+++ src/TianWen.Lib/Imaging/Stacking/InRamAllFramesStrategy.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace TianWen.Lib.Imaging.Calibration;
+namespace TianWen.Lib.Imaging.Stacking;
 
 /// <summary>
 /// Highest-fidelity strategy: warp every frame into memory, run the
