+++ src/TianWen.Lib/Imaging/Stacking/IntegrationStrategySelector.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
 
-namespace TianWen.Lib.Imaging.Calibration;
+namespace TianWen.Lib.Imaging.Stacking;
 
 /// <summary>
 /// Two-phase picker over <see cref="IIntegrationStrategy"/>:
