+++ src/TianWen.Lib/Imaging/Stacking/Integrator.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace TianWen.Lib.Imaging.Calibration;
+namespace TianWen.Lib.Imaging.Stacking;
 
 /// <summary>
 /// Options for the stack <see cref="Integrator"/>: rejector + combiner +
