+++ src/TianWen.Lib/Imaging/Stacking/LiveAccumulatorStrategy.cs
@@ -2,7 +2,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace TianWen.Lib.Imaging.Calibration;
+namespace TianWen.Lib.Imaging.Stacking;
 
 /// <summary>
 /// Frame-at-a-time online accumulator (Welford mean + variance) for live
