+++ src/TianWen.Lib/Imaging/Stacking/MemoryMappedFitsSink.cs
@@ -3,7 +3,7 @@ using System.IO;
 using System.IO.MemoryMappedFiles;
 using System.Runtime.InteropServices;
 
-namespace TianWen.Lib.Imaging.Calibration;
+namespace TianWen.Lib.Imaging.Stacking;
 
 /// <summary>
 /// Disk-backed <see cref="IIntegrationSink"/>: stores the master canvas in
