+++ src/TianWen.Lib/Imaging/Stacking/StreamingFrameStaging.cs
@@ -4,7 +4,7 @@ using System.Drawing;
 using System.IO;
 using System.Runtime.InteropServices;
 
-namespace TianWen.Lib.Imaging.Calibration;
+namespace TianWen.Lib.Imaging.Stacking;
 
 /// <summary>
 /// Disk-backed staging for warped frames during stacking. Header byte 12 is
