+++ src/TianWen.Lib/Imaging/Stacking/StreamingIntegrator.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace TianWen.Lib.Imaging.Calibration;
+namespace TianWen.Lib.Imaging.Stacking;
 
 /// <summary>
 /// A staged aligned frame for the <see cref="StreamingIntegrator"/>. Owns a
